Move meta to new module
