+++ saleor/graphql/account/mutations/account.py
@@ -10,14 +10,10 @@ from ....checkout import AddressType
 from ....core.utils.url import validate_storefront_url
 from ...account.enums import AddressTypeEnum
 from ...account.types import Address, AddressInput, User
-from ...core.mutations import (
-    BaseMutation,
-    ModelDeleteMutation,
-    ModelMutation,
-    UpdateMetaBaseMutation,
-)
-from ...core.types import MetaInput
+from ...core.mutations import BaseMutation, ModelDeleteMutation, ModelMutation
 from ...core.types.common import AccountError
+from ...meta.mutations import UpdateMetaBaseMutation
+from ...meta.types import MetaInput
 from .base import (
     INVALID_TOKEN,
     BaseAddressDelete,
