+++ saleor/graphql/core/types/__init__.py
@@ -10,7 +10,6 @@ from .common import (
     Weight,
 )
 from .filter_input import FilterInputObjectType
-from .meta import MetaInput, MetaPath, ObjectWithMetadata
 from .money import VAT, Money, MoneyRange, ReducedRate, TaxedMoney, TaxedMoneyRange
 from .sort_input import SortInputObjectType
 from .upload import Upload
