+++ saleor/graphql/order/mutations/orders.py
@@ -18,14 +18,11 @@ from ....order.error_codes import OrderErrorCode
 from ....order.utils import get_valid_shipping_methods_for_order
 from ....payment import CustomPaymentChoices, PaymentError, gateway
 from ...account.types import AddressInput
-from ...core.mutations import (
-    BaseMutation,
-    ClearMetaBaseMutation,
-    UpdateMetaBaseMutation,
-)
+from ...core.mutations import BaseMutation
 from ...core.scalars import Decimal
-from ...core.types import MetaInput, MetaPath
 from ...core.types.common import OrderError
+from ...meta.mutations import ClearMetaBaseMutation, UpdateMetaBaseMutation
+from ...meta.types import MetaInput, MetaPath
 from ...order.mutations.draft_orders import DraftOrderUpdate
 from ...order.types import Order, OrderEvent
 from ...shipping.types import ShippingMethod
